@@ -1,3 +1,4 @@
+///<reference path="../model/dataset.ts"/>
 import { Dispatch } from 'redux';
 
 import { AppState } from '../states/appState';
@@ -8,16 +9,24 @@ import {
     RemoveActivity,
     removeActivity,
     SelectDataset,
-    selectDataset
+    selectDataset,
 } from './controlActions';
-import { Dataset } from '../model/dataset';
+import {
+    Dataset,
+} from '../model/dataset';
 import { TimeSeries } from '../model/timeSeries';
 import { ColorBars } from '../model/colorBar';
 import { I18N } from '../config';
-import { selectedServerSelector } from '../selectors/controlSelectors';
+import {
+    selectedDatasetIdSelector, selectedDatasetTimeDimensionSelector,
+    selectedDatasetVariableSelector, selectedPlaceIdSelector, selectedPlaceSelector,
+    selectedServerSelector
+} from '../selectors/controlSelectors';
 import { Server } from '../model/server';
 import { MessageLogAction, postMessage } from './messageLogActions';
-import { PlaceGroup } from '../model/place';
+import { findPlaceInPlaceGroups, Place, PlaceGroup } from '../model/place';
+import * as geojson from 'geojson';
+import { placeGroupsSelector } from '../selectors/dataSelectors';
 
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -73,34 +82,126 @@ export function updateDatasetPlaceGroup(datasetId: string,
     return {type: UPDATE_DATASET_PLACE_GROUP, datasetId, placeGroup};
 }
 
-
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-export const UPDATE_COLOR_BARS = 'UPDATE_COLOR_BARS';
-export type UPDATE_COLOR_BARS = typeof UPDATE_COLOR_BARS;
+export const ADD_USER_PLACE = 'ADD_USER_PLACE';
+export type ADD_USER_PLACE = typeof ADD_USER_PLACE;
 
-export interface UpdateColorBars {
-    type: UPDATE_COLOR_BARS;
-    colorBars: ColorBars;
+export interface AddUserPlace {
+    type: ADD_USER_PLACE;
+    id: string;
+    label: string;
+    color: string;
+    geometry: geojson.Geometry;
 }
 
-export function updateColorBars() {
-    return (dispatch: Dispatch<UpdateColorBars | MessageLogAction>, getState: () => AppState) => {
-        const apiServer = selectedServerSelector(getState());
-
-        api.getColorBars(apiServer.url)
-           .then((colorBars: ColorBars) => {
-               dispatch(_updateColorBars(colorBars));
-           })
-           .catch(error => {
-               dispatch(postMessage('error', error.message || `${error}`));
-           });
+export function addUserPlace(id: string, label: string, color: string, geometry: geojson.Geometry) {
+    return (dispatch: Dispatch<AddUserPlace>, getState: () => AppState) => {
+        dispatch(_addUserPlace(id, label, color, geometry));
+        if (getState().controlState.autoShowTimeSeries) {
+            dispatch(addTimeSeries() as any);
+        }
     };
 }
 
-export function _updateColorBars(colorBars: ColorBars): UpdateColorBars {
-    return {type: UPDATE_COLOR_BARS, colorBars};
+export function _addUserPlace(id: string, label: string, color: string, geometry: geojson.Geometry): AddUserPlace {
+    return {type: ADD_USER_PLACE, id, label, color, geometry};
 }
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+export const REMOVE_USER_PLACE = 'REMOVE_USER_PLACE';
+export type REMOVE_USER_PLACE = typeof REMOVE_USER_PLACE;
+
+export interface RemoveUserPlace {
+    type: REMOVE_USER_PLACE;
+    id: string;
+    places: Place[];
+}
+
+export function removeUserPlace(id: string, places: Place[]): RemoveUserPlace {
+    return {type: REMOVE_USER_PLACE, id, places};
+}
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+export const REMOVE_ALL_USER_PLACES = 'REMOVE_ALL_USER_PLACES';
+export type REMOVE_ALL_USER_PLACES = typeof REMOVE_ALL_USER_PLACES;
+
+export interface RemoveAllUserPlaces {
+    type: REMOVE_ALL_USER_PLACES;
+}
+
+export function removeAllUserPlaces(): RemoveAllUserPlaces {
+    return {type: REMOVE_ALL_USER_PLACES};
+}
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+export function addTimeSeries() {
+    return (dispatch: Dispatch<UpdateTimeSeries | MessageLogAction>, getState: () => AppState) => {
+        const apiServer = selectedServerSelector(getState());
+
+        const selectedDatasetId = selectedDatasetIdSelector(getState());
+        const selectedDatasetTimeDim = selectedDatasetTimeDimensionSelector(getState());
+        const selectedVariable = selectedDatasetVariableSelector(getState());
+        const selectedPlaceId = selectedPlaceIdSelector(getState());
+        const selectedPlace = selectedPlaceSelector(getState())!;
+        const timeSeriesUpdateMode = getState().controlState.timeSeriesUpdateMode;
+        const inclStDev = getState().controlState.showTimeSeriesErrorBars;
+
+        const placeGroups = placeGroupsSelector(getState());
+
+        if (selectedDatasetId && selectedVariable && selectedPlaceId && selectedDatasetTimeDim) {
+            const timeLabels = selectedDatasetTimeDim.labels;
+            const numTimeLabels = timeLabels.length;
+            const timeChunkSize = 16;
+            let endTimeIndex = numTimeLabels - 1;
+            let startTimeIndex = endTimeIndex - timeChunkSize + 1;
+
+            const getTimeSeriesChunk = () => {
+                const startDateLabel = startTimeIndex >= 0 ? timeLabels[startTimeIndex] : null;
+                const endDateLabel = timeLabels[endTimeIndex];
+                return api.getTimeSeriesForGeometry(apiServer.url,
+                                                    selectedDatasetId,
+                                                    selectedVariable,
+                                                    selectedPlace.id,
+                                                    selectedPlace.geometry,
+                                                    startDateLabel,
+                                                    endDateLabel,
+                                                    inclStDev);
+            };
+
+            const successAction = (timeSeries: TimeSeries | null) => {
+                if (timeSeries !== null && isValidPlace(placeGroups, selectedPlace.id)) {
+                    const hasMore = startTimeIndex > 0;
+                    const dataProgress = hasMore ? (numTimeLabels - startTimeIndex) / numTimeLabels : 1.0;
+                    dispatch(updateTimeSeries({...timeSeries, dataProgress},
+                                              timeSeriesUpdateMode,
+                                              endTimeIndex === numTimeLabels - 1 ? 'new' : 'append'));
+                    if (hasMore && isValidPlace(placeGroups, selectedPlace.id)) {
+                        startTimeIndex -= timeChunkSize;
+                        endTimeIndex -= timeChunkSize;
+                        getTimeSeriesChunk().then(successAction);
+                    }
+                } else {
+                    dispatch(postMessage('info', 'No data found here'));
+                }
+            };
+
+            getTimeSeriesChunk()
+                .then(successAction)
+                .catch((error: any) => {
+                    dispatch(postMessage('error', error + ''));
+                });
+        }
+    };
+}
+
+function isValidPlace(placeGroups: PlaceGroup[], placeId: string) {
+    return findPlaceInPlaceGroups(placeGroups, placeId) !== null;
+}
+
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -110,11 +211,11 @@ export type UPDATE_TIME_SERIES = typeof UPDATE_TIME_SERIES;
 export interface UpdateTimeSeries {
     type: UPDATE_TIME_SERIES;
     timeSeries: TimeSeries;
-    updateMode: 'add' | 'replace';
+    updateMode: 'add' | 'replace' | 'remove';
     dataMode: 'new' | 'append';
 }
 
-export function updateTimeSeries(timeSeries: TimeSeries, updateMode: 'add' | 'replace', dataMode: 'new' | 'append'): UpdateTimeSeries {
+export function updateTimeSeries(timeSeries: TimeSeries, updateMode: 'add' | 'replace' | 'remove', dataMode: 'new' | 'append'): UpdateTimeSeries {
     return {type: UPDATE_TIME_SERIES, timeSeries, updateMode, dataMode};
 }
 
@@ -174,13 +275,45 @@ export function _configureServers(servers: Server[], selectedServerId: string): 
     return {type: CONFIGURE_SERVERS, servers, selectedServerId};
 }
 
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+export const UPDATE_COLOR_BARS = 'UPDATE_COLOR_BARS';
+export type UPDATE_COLOR_BARS = typeof UPDATE_COLOR_BARS;
+
+export interface UpdateColorBars {
+    type: UPDATE_COLOR_BARS;
+    colorBars: ColorBars;
+}
+
+export function updateColorBars() {
+    return (dispatch: Dispatch<UpdateColorBars | MessageLogAction>, getState: () => AppState) => {
+        const apiServer = selectedServerSelector(getState());
+
+        api.getColorBars(apiServer.url)
+           .then((colorBars: ColorBars) => {
+               dispatch(_updateColorBars(colorBars));
+           })
+           .catch(error => {
+               dispatch(postMessage('error', error.message || `${error}`));
+           });
+    };
+}
+
+export function _updateColorBars(colorBars: ColorBars): UpdateColorBars {
+    return {type: UPDATE_COLOR_BARS, colorBars};
+}
+
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 export type DataAction =
     UpdateDatasets
     | UpdateDatasetPlaceGroup
-    | UpdateColorBars
+    | AddUserPlace
+    | RemoveUserPlace
+    | RemoveAllUserPlaces
     | UpdateTimeSeries
     | RemoveTimeSeriesGroup
     | RemoveAllTimeSeries
-    | ConfigureServers;
+    | ConfigureServers
+    | UpdateColorBars;
