@@ -3,35 +3,32 @@ import { withStyles, WithStyles, createStyles, Theme } from '@material-ui/core/s
 import Input from '@material-ui/core/Input';
 import InputLabel from '@material-ui/core/InputLabel';
 import MenuItem from '@material-ui/core/MenuItem';
-import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
-import Checkbox from "@material-ui/core/Checkbox";
-import ListItemText from "@material-ui/core/ListItemText";
+import Checkbox from '@material-ui/core/Checkbox';
+import ListItemText from '@material-ui/core/ListItemText';
+import IconButton from '@material-ui/core/IconButton';
+import RemoveCircleOutlineIcon from '@material-ui/icons/RemoveCircleOutline';
 
-import { Dataset } from '../model/dataset';
 import { PlaceGroup } from '../model/place';
-import { WithLocale } from "../util/lang";
-import { I18N } from "../config";
+import { WithLocale } from '../util/lang';
+import { I18N } from '../config';
+import ControlBarItem from './ControlBarItem';
 
 
+// noinspection JSUnusedLocalSymbols
 const styles = (theme: Theme) => createStyles(
     {
-        formControl: {
-            marginRight: theme.spacing.unit * 2,
-            marginBottom: theme.spacing.unit,
-            minWidth: 120,
-        },
-        selectEmpty: {
-            marginTop: theme.spacing.unit * 2,
+        select: {
+            minWidth: '5em',
         },
     });
 
 interface PlaceGroupSelectProps extends WithStyles<typeof styles>, WithLocale {
-    datasets: Dataset[];
-
     selectedPlaceGroupIds: string[] | null;
     placeGroups: PlaceGroup[];
-    selectPlaceGroups: (placeGroupIds: string[] | null, dataset: Dataset[]) => void;
+    selectPlaceGroups: (placeGroupIds: string[] | null) => void;
+    removeAllUserPlaces: () => void;
+
     selectedPlaceGroupsTitle: string;
 
     language?: string;
@@ -40,8 +37,13 @@ interface PlaceGroupSelectProps extends WithStyles<typeof styles>, WithLocale {
 class PlaceGroupsSelect extends React.Component<PlaceGroupSelectProps> {
 
     handlePlaceGroupsChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-        console.log("handlePlaceGroupsChange: ", event.target.value);
-        this.props.selectPlaceGroups(event.target.value as any as string[] || null, this.props.datasets);
+        const {selectPlaceGroups} = this.props;
+        selectPlaceGroups(event.target.value as any as string[] || null);
+    };
+
+    handleRemoveButtonClick = () => {
+        const {removeAllUserPlaces} = this.props;
+        removeAllUserPlaces();
     };
 
     renderSelectedPlaceGroupsTitle = () => {
@@ -54,38 +56,58 @@ class PlaceGroupsSelect extends React.Component<PlaceGroupSelectProps> {
         const placeGroups = this.props.placeGroups || [];
         const selectedPlaceGroupIds = this.props.selectedPlaceGroupIds || [];
 
-        if (placeGroups.length === 0) {
+        if (placeGroups.length == 0 || placeGroups.length == 1 && placeGroups[0].id == 'user') {
             return null;
         }
 
-        return (
+        const placeGroupsSelectLabel = (
+            <InputLabel
+                shrink
+                htmlFor="place-groups-select"
+            >
+                {I18N.get('Places')}
+            </InputLabel>
+        );
 
-            <FormControl className={classes.formControl}>
-                <InputLabel
-                    shrink
-                    htmlFor="place-groups-select"
-                >
-                    {I18N.get("Places")}
-                </InputLabel>
-                <Select
-                    multiple
-                    onChange={this.handlePlaceGroupsChange}
-                    input={<Input name="place-groups" id="place-groups-select"/>}
-                    value={selectedPlaceGroupIds}
-                    renderValue={this.renderSelectedPlaceGroupsTitle}
-                    name="place-groups"
-                >
-                    {placeGroups.map(placeGroup => (
-                        <MenuItem
-                            key={placeGroup.id}
-                            value={placeGroup.id}
-                        >
-                            <Checkbox checked={selectedPlaceGroupIds.indexOf(placeGroup.id) > -1}/>
-                            <ListItemText primary={placeGroup.title}/>
-                        </MenuItem>
-                    ))}
-                </Select>
-            </FormControl>
+        const placeGroupsSelect = (
+            <Select
+                multiple
+                displayEmpty
+                onChange={this.handlePlaceGroupsChange}
+                input={<Input name="place-groups" id="place-groups-select"/>}
+                value={selectedPlaceGroupIds}
+                renderValue={this.renderSelectedPlaceGroupsTitle}
+                name="place-groups"
+                className={classes.select}
+            >
+                {placeGroups.map(placeGroup => (
+                    <MenuItem
+                        key={placeGroup.id}
+                        value={placeGroup.id}
+                    >
+                        <Checkbox checked={selectedPlaceGroupIds.indexOf(placeGroup.id) > -1}/>
+                        <ListItemText primary={placeGroup.title}/>
+                    </MenuItem>
+                ))}
+            </Select>
+        );
+
+        const removeEnabled = selectedPlaceGroupIds.length === 1 && selectedPlaceGroupIds[0] === 'user';
+        const placeGroupRemoveButton = (
+            <IconButton
+                disabled={!removeEnabled}
+                onClick={this.handleRemoveButtonClick}
+            >
+                {<RemoveCircleOutlineIcon/>}
+            </IconButton>
+        );
+
+        return (
+            <ControlBarItem
+                label={placeGroupsSelectLabel}
+                control={placeGroupsSelect}
+                actions={placeGroupRemoveButton}
+            />
         );
     }
 }
