@@ -10,15 +10,16 @@ import {
     Legend,
     AxisDomain,
     TooltipPayload,
-    ReferenceArea, ReferenceLine, TooltipProps, ErrorBar
+    ReferenceArea, ReferenceLine, TooltipProps, ErrorBar, DotProps
 } from 'recharts';
 import { Theme, createStyles, withStyles, WithStyles } from '@material-ui/core/styles';
 import IconButton from '@material-ui/core/IconButton';
-import ZoomOutMap from '@material-ui/icons/ZoomOutMap';
-import Close from '@material-ui/icons/Close';
+import AllOutIcon from '@material-ui/icons/AllOut';
+import CloseIcon from '@material-ui/icons/Close';
 import Typography from '@material-ui/core/Typography';
+import { CircularProgress } from "@material-ui/core";
 
-import { equalTimeRanges, Time, TimeRange, TimeSeriesGroup, TimeSeriesPoint } from '../model/timeSeries';
+import { equalTimeRanges, Time, TimeRange, TimeSeries, TimeSeriesGroup, TimeSeriesPoint } from '../model/timeSeries';
 import { utcTimeToLocalDateString, utcTimeToLocalDateTimeString } from '../util/time';
 import {
     I18N,
@@ -26,28 +27,42 @@ import {
     LINE_CHART_STROKE_SHADE_LIGHT_THEME,
     USER_PLACES_COLORS
 } from '../config';
-import { WithLocale } from "../util/lang";
+import { WithLocale } from '../util/lang';
+import { Place, PlaceInfo } from '../model/place';
 
 
 const styles = (theme: Theme) => createStyles(
     {
         chartContainer: {
-            userSelect: 'none',
-            position: 'relative',
-            width: "99%",
-            height: '40vh',
+            // userSelect: 'none',
+            marginTop: theme.spacing(1),
+            width: '99%',
+            height: '32vh',
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'flex-stretch',
+        },
+        responsiveContainer: {
+            flexGrow: 1,
         },
         zoomOutButton: {
             position: 'absolute',
-            right: 8 * theme.spacing.unit,
-            margin: theme.spacing.unit,
+            right: 8 * theme.spacing(1),
+            margin: theme.spacing(1),
             zIndex: 1000,
             opacity: 0.8,
         },
         removeTimeSeriesGroup: {
             position: 'absolute',
-            right: theme.spacing.unit,
-            margin: theme.spacing.unit,
+            right: theme.spacing(1),
+            margin: theme.spacing(1),
+            zIndex: 1000,
+            opacity: 0.8,
+        },
+        removeTimeSeriesProgress: {
+            position: 'absolute',
+            right: theme.spacing(1),
+            margin: theme.spacing(2.5),
             zIndex: 1000,
             opacity: 0.8,
         },
@@ -56,15 +71,15 @@ const styles = (theme: Theme) => createStyles(
             opacity: 0.8,
             color: 'white',
             border: '2px solid black',
-            borderRadius: theme.spacing.unit * 2,
-            padding: theme.spacing.unit * 1.5,
+            borderRadius: theme.spacing(2),
+            padding: theme.spacing(1.5),
         },
         toolTipValue: {
             fontWeight: 'bold',
         },
         toolTipLabel: {
             fontWeight: 'bold',
-            paddingBottom: theme.spacing.unit,
+            paddingBottom: theme.spacing(1),
         },
         chartTitle: {}
     });
@@ -80,7 +95,18 @@ interface TimeSeriesChartProps extends WithStyles<typeof styles>, WithLocale {
     selectedTimeRange?: TimeRange | null;
     selectTimeRange?: (timeRange: TimeRange | null) => void;
 
+    showPointsOnly: boolean;
+    showErrorBars: boolean;
+
+    selectTimeSeries?: (timeSeriesGroupId: string, timeSeriesIndex: number, timeSeries: TimeSeries) => void;
+
     removeTimeSeriesGroup?: (id: string) => void;
+    completed: number[];
+
+    placeInfos?: { [placeId: string]: PlaceInfo };
+
+    selectPlace: (placeId: string | null, places: Place[], showInMap: boolean) => void;
+    places: Place[];
 }
 
 interface TimeSeriesChartState {
@@ -102,7 +128,10 @@ class TimeSeriesChart extends React.Component<TimeSeriesChartProps, TimeSeriesCh
     }
 
     render() {
-        const {classes, timeSeriesGroup, selectedTime, selectedTimeRange, dataTimeRange, theme} = this.props;
+        const {
+            classes, timeSeriesGroup, selectedTime, selectedTimeRange,
+            dataTimeRange, theme, placeInfos, showErrorBars, showPointsOnly,
+        } = this.props;
 
         const strokeShade = theme.palette.type === 'light' ? LINE_CHART_STROKE_SHADE_LIGHT_THEME : LINE_CHART_STROKE_SHADE_DARK_THEME;
         const lightStroke = theme.palette.primary.light;
@@ -118,6 +147,25 @@ class TimeSeriesChart extends React.Component<TimeSeriesChartProps, TimeSeriesCh
         }
 
         const lines = timeSeriesGroup.timeSeriesArray.map((ts, i) => {
+
+            const source = ts.source;
+            let lineName = source.variableName;
+            let lineColor = 'yellow';
+            if (placeInfos) {
+                const placeInfo = placeInfos[source.placeId];
+                if (placeInfo) {
+                    const {place, label, color} = placeInfo;
+                    if (place.geometry.type === 'Point') {
+                        const lon = place.geometry.coordinates[0];
+                        const lat = place.geometry.coordinates[1];
+                        lineName += ` (${label}: ${lat.toFixed(5)},${lon.toFixed(5)})`;
+                    } else {
+                        lineName += ` (${label})`;
+                    }
+                    lineColor = color;
+                }
+            }
+
             const data: TimeSeriesPoint[] = [];
             let hasErrorBars = false;
             ts.data.forEach(point => {
@@ -132,39 +180,39 @@ class TimeSeriesChart extends React.Component<TimeSeriesChartProps, TimeSeriesCh
                     }
                     if (time1Ok && time2Ok) {
                         data.push(point);
-                    }
-                    // noinspection SuspiciousTypeOfGuard
-                    if ((typeof point.uncertainty) === "number") {
-                        hasErrorBars = true;
+                        // noinspection SuspiciousTypeOfGuard
+                        if ((typeof point.uncertainty) === 'number') {
+                            hasErrorBars = true;
+                        }
                     }
                 }
             });
+            const shadedLineColor = USER_PLACES_COLORS[lineColor][strokeShade];
             let errorBar;
-            if (hasErrorBars) {
+            if (showErrorBars && hasErrorBars) {
                 errorBar = (
                     <ErrorBar
                         dataKey="uncertainty"
                         width={4}
-                        strokeWidth={2}
-                        stroke={USER_PLACES_COLORS[ts.color][strokeShade]}
+                        strokeWidth={1}
+                        stroke={shadedLineColor}
                     />
                 );
             }
-            const source = ts.source;
             return (
                 <Line
                     key={i}
                     type="monotone"
-                    name={source.variableName}
+                    name={lineName}
                     unit={source.variableUnits}
                     data={data}
                     dataKey="average"
-                    connectNulls={true}
-                    dot={true}
-                    stroke={USER_PLACES_COLORS[ts.color][strokeShade]}
+                    dot={<CustomizedDot radius={4} stroke={shadedLineColor} fill={'white'} strokeWidth={3}/>}
+                    activeDot={<CustomizedDot radius={4} stroke={'white'} fill={shadedLineColor} strokeWidth={3}/>}
+                    stroke={showPointsOnly ? '#00000000' :shadedLineColor}
                     strokeWidth={3 * (ts.dataProgress || 1)}
-                    activeDot={true}
                     isAnimationActive={ts.dataProgress === 1.0}
+                    onClick={() => this.handleTimeSeriesClick(timeSeriesGroup.id, i, ts)}
                 >{errorBar}</Line>
             );
         });
@@ -193,26 +241,35 @@ class TimeSeriesChart extends React.Component<TimeSeriesChartProps, TimeSeriesCh
                     aria-label="Zoom Out"
                     onClick={this.handleZoomOutButtonClick}
                 >
-                    <ZoomOutMap/>
+                    <AllOutIcon/>
                 </IconButton>
             );
             actionButtons.push(zoomOutButton);
         }
+        const progress = this.props.completed.reduce((a: number, b: number) => a + b , 0) / this.props.completed.length;
+        const loading = !!(progress > 0 && progress < 100);
 
-        const removeAllButton = (
-            <IconButton
-                key={'removeTimeSeriesGroup'}
-                className={classes.removeTimeSeriesGroup}
-                aria-label="Close"
-                onClick={this.handleRemoveTimeSeriesGroupClick}
-            >
-                <Close/>
-            </IconButton>
+        //const progressBar = (<LinearProgress className={classes.removeTimeSeriesGroup} color="secondary" variant="determinate" value={progress}/>);
+        const progressBar = (<CircularProgress size={24} className={classes.removeTimeSeriesProgress} color={"secondary"}/>);
+
+        const removeButton = (
+            (
+                    <IconButton
+                        key={'removeTimeSeriesGroup'}
+                        className={classes.removeTimeSeriesGroup}
+                        aria-label="Close"
+                        onClick={this.handleRemoveTimeSeriesGroupClick}
+                    >
+                        <CloseIcon/>
+                    </IconButton>)
         );
+
+        const removeAllButton = loading ? progressBar : removeButton;
+
         actionButtons.push(removeAllButton);
 
-        const timeSeriesText = I18N.get("Time-Series");
-        const unitsText = timeSeriesGroup.variableUnits || I18N.get("unknown units");
+        const timeSeriesText = I18N.get('Time-Series');
+        const unitsText = timeSeriesGroup.variableUnits || I18N.get('unknown units');
         const chartTitle = `${timeSeriesText} (${unitsText})`;
 
         // 99% per https://github.com/recharts/recharts/issues/172
@@ -220,7 +277,7 @@ class TimeSeriesChart extends React.Component<TimeSeriesChartProps, TimeSeriesCh
             <div className={classes.chartContainer}>
                 <Typography className={classes.chartTitle}>{chartTitle}</Typography>
                 {actionButtons}
-                <ResponsiveContainer width="99%" height={320}>
+                <ResponsiveContainer width="99%" className={classes.responsiveContainer}>
                     <LineChart onMouseDown={this.handleMouseDown}
                                onMouseMove={this.handleMouseMove}
                                onMouseUp={this.handleMouseUp}
@@ -267,6 +324,15 @@ class TimeSeriesChart extends React.Component<TimeSeriesChartProps, TimeSeriesCh
         this.setState(TimeSeriesChart.clearState());
     };
 
+    readonly handleTimeSeriesClick = (timeSeriesGroupId: string, timeSeriesIndex: number, timeSeries: TimeSeries) => {
+        const {selectTimeSeries, selectPlace, places} = this.props;
+        console.log('handleTimeSeriesClick:', timeSeriesGroupId, timeSeriesIndex, timeSeries);
+        if (!!selectTimeSeries) {
+            selectTimeSeries(timeSeriesGroupId, timeSeriesIndex, timeSeries);
+        }
+        selectPlace(timeSeries.source.placeId, places, true);
+    };
+
     readonly handleMouseDown = (event: any) => {
         if (event) {
             this.setState(TimeSeriesChart.newState(false, event.activeLabel, null));
@@ -291,6 +357,9 @@ class TimeSeriesChart extends React.Component<TimeSeriesChartProps, TimeSeriesCh
     readonly handleRemoveTimeSeriesGroupClick = () => {
         if (this.props.removeTimeSeriesGroup) {
             this.props.removeTimeSeriesGroup(this.props.timeSeriesGroup.id);
+            this.setState(TimeSeriesChart.newState(this.state.isDragging,
+                this.state.firstTime,
+                this.state.secondTime))
         }
     };
 
@@ -319,7 +388,8 @@ class TimeSeriesChart extends React.Component<TimeSeriesChartProps, TimeSeriesCh
         });
     };
 
-    private static newState(isDragging: boolean, firstTime: number | null, secondTime: number | null): TimeSeriesChartState {
+    private static newState(isDragging: boolean, firstTime: number | null, secondTime: number | null):
+        TimeSeriesChartState {
         return {isDragging, firstTime, secondTime};
     }
 
@@ -382,3 +452,34 @@ class _CustomTooltip extends React.PureComponent<_CustomTooltipProps> {
 }
 
 const CustomTooltip = withStyles(styles)(_CustomTooltip);
+
+interface CustomizedDotProps extends DotProps {
+    radius: number;
+    stroke: string;
+    strokeWidth: number;
+    fill: string;
+}
+
+
+const CustomizedDot = (props: CustomizedDotProps) => {
+
+    const {cx, cy, radius, stroke, fill, strokeWidth} = props;
+
+    const vpSize = 1024;
+    const totalRadius = radius + 0.5 * strokeWidth;
+    const totalDiameter = 2 * totalRadius;
+
+    const r = Math.floor(100 * radius / totalDiameter + 0.5) + '%';
+    const sw = Math.floor(100 * strokeWidth / totalDiameter + 0.5) + '%';
+
+    // noinspection SuspiciousTypeOfGuard
+    if (typeof cx === 'number' && typeof cy === 'number') {
+        return (
+            <svg x={cx - totalRadius} y={cy - totalRadius} width={totalDiameter} height={totalDiameter}
+                 viewBox={`0 0 ${vpSize} ${vpSize}`}>
+                <circle cx='50%' cy='50%' r={r} strokeWidth={sw} stroke={stroke} fill={fill}/>
+            </svg>
+        );
+    }
+    return null;
+};
