@@ -21,6 +21,7 @@ interface MapProps extends ol.olx.MapOptions {
     mapObjects?: { [id: string]: ol.Object };
     onClick?: (event: ol.MapBrowserEvent) => void;
     onMapRef?: (map: ol.Map | null) => void;
+    isStale?: boolean;
 }
 
 interface MapState {
@@ -28,20 +29,25 @@ interface MapState {
 
 const DEFAULT_CONTAINER_SYTLE: React.CSSProperties = {height: '100%'};
 
-const DEFAULT_VIEW = new ol.View(
-    {
-        center: ol.proj.fromLonLat([0, 0]),
-        zoom: 2
-    }
-);
-
 export class Map extends React.Component<MapProps, MapState> {
 
-    private contextValue: MapContext;
+    private readonly contextValue: MapContext;
 
     constructor(props: MapProps) {
         super(props);
-        this.contextValue = {mapObjects: this.props.mapObjects || {}};
+        // console.log("Map.constructor: id =", this.props.id);
+
+        const {id, mapObjects} = props;
+        if (mapObjects) {
+            this.contextValue = {
+                map: mapObjects[id] as ol.Map || undefined,
+                mapObjects: mapObjects
+            };
+        } else {
+            this.contextValue = {
+                mapObjects: {}
+            };
+        }
     }
 
     private getMapOptions(): ol.olx.MapOptions {
@@ -58,19 +64,81 @@ export class Map extends React.Component<MapProps, MapState> {
         }
     };
 
-    componentDidMount(): void {
-        // console.log("Map.componentDidMount: new Map!");
-        const mapOptions = this.getMapOptions();
-        const target = this.contextValue.mapDiv!;
-        const map = new ol.Map({view: DEFAULT_VIEW, ...mapOptions, target});
-        map.set("objectId", this.props.id);
-        this.contextValue.map = map;
-        this.contextValue.mapObjects[this.props.id] = map;
+    private handleRef = (mapDiv: HTMLDivElement | null) => {
+        this.contextValue.mapDiv = mapDiv;
+    };
 
+    private handleResize = () => {
+        const mapDiv = this.contextValue.mapDiv;
+        const map = this.contextValue.map;
+        if (mapDiv && map) {
+            map.updateSize();
+            const view = map.getView();
+            const minZoom = this.getMinZoom(mapDiv);
+            if (minZoom !== view.getMinZoom()) {
+                view.setMinZoom(minZoom);
+            }
+        }
+    };
+
+    private getMinZoom = (target: HTMLDivElement) => {
+        // Adjust the view's minZoom so there is only one world,
+        // see https://openlayers.org/en/latest/examples/min-zoom.html
+        const size = target.clientWidth;
+        const minZoom = Math.LOG2E * Math.log(size / 256);
+        if (minZoom >= 0.0) {
+            return minZoom;
+        }
+        return 0;
+    };
+
+    componentDidMount(): void {
+        // console.log('Map.componentDidMount: id =', this.props.id);
+
+        const {id} = this.props;
+        const mapDiv = this.contextValue.mapDiv!;
+
+        let map: ol.Map | undefined;
+        if (this.props.isStale) {
+            const mapObject = this.contextValue.mapObjects[id];
+            if (mapObject && mapObject['addControl'] && mapObject['addLayer'] && mapObject['setTarget']) {
+                map = mapObject as ol.Map;
+                map.setTarget(mapDiv);
+            }
+        }
+
+        if (!map) {
+            const initialZoom = this.getMinZoom(mapDiv);
+            const view = new ol.View({
+                                         center: ol.proj.fromLonLat([0, 0]),
+                                         minZoom: initialZoom,
+                                         zoom: initialZoom,
+                                     });
+            map = new ol.Map({
+                                 view,
+                                 ...this.getMapOptions(),
+                                 target: mapDiv
+                             });
+        }
+
+        this.contextValue.map = map;
+        this.contextValue.mapObjects[id] = map;
+
+        map.set('objectId', this.props.id);
         map.on('click', this.handleClick);
+        map.updateSize();
 
         // Force update so we can pass this.map as context to all children in next render()
         this.forceUpdate();
+
+        // Add resize listener so we can adjust the view's minZoom.
+        // See https://openlayers.org/en/latest/examples/min-zoom.html
+        window.addEventListener('resize', this.handleResize);
+        mapDiv.onresize = () => {
+            if (this.contextValue.map) {
+                this.contextValue.map.updateSize();
+            }
+        };
 
         const onMapRef = this.props.onMapRef;
         if (onMapRef) {
@@ -79,19 +147,29 @@ export class Map extends React.Component<MapProps, MapState> {
     }
 
     componentDidUpdate(prevProps: Readonly<MapProps>): void {
-        // console.log("Map.componentDidUpdate: update Map!");
-        const map = this.contextValue.map;
+        // console.log('Map.componentDidUpdate: id =', this.props.id);
+
+        const map = this.contextValue.map!;
+        const mapDiv = this.contextValue.mapDiv!;
         const mapOptions = this.getMapOptions();
-        map!.setProperties({...mapOptions, target: this.contextValue.mapDiv});
+        map.setProperties({...mapOptions});
+        map.setTarget(mapDiv);
+        map.updateSize();
     }
 
     componentWillUnmount(): void {
-        // console.log("Map.componentWillUnmount: remove Map!");
+        // console.log('Map.componentWillUnmount: id =', this.props.id);
+
+        const mapDiv = this.contextValue.mapDiv!;
+        mapDiv.onresize = null;
+
+        // Remove resize listener so we can adjust the view's minZoom.
+        window.removeEventListener('resize', this.handleResize);
+
         const onMapRef = this.props.onMapRef;
         if (onMapRef) {
             onMapRef(null);
         }
-        this.contextValue = {mapObjects: {}};
     }
 
     render() {
@@ -109,17 +187,5 @@ export class Map extends React.Component<MapProps, MapState> {
             </div>
         );
     }
-
-    private handleRef = (mapDiv: HTMLDivElement | null) => {
-        if (mapDiv !== null) {
-            mapDiv.onresize = () => {
-                const map = this.contextValue.map;
-                if (map) {
-                    map.updateSize();
-                }
-            };
-        }
-        this.contextValue.mapDiv = mapDiv;
-    };
 }
 
