@@ -7,9 +7,10 @@ import DatasetSelect from './DatasetSelect';
 import VariableSelect from './VariableSelect';
 import PlaceGroupsSelect from './PlaceGroupsSelect';
 import PlaceSelect from './PlaceSelect';
+import MapInteractionsBar from "./MapInteractionsBar";
 import TimeSelect from './TimeSelect';
 import TimeSlider from './TimeSlider';
-import TimeSeriesModeSelect from './TimeSeriesModeSelect';
+import TimePlayer from "./TimePlayer";
 
 
 const mapStateToProps = (state: AppState) => {
@@ -28,9 +29,10 @@ class _ControlBar extends React.Component {
                 <VariableSelect/>
                 <PlaceGroupsSelect/>
                 <PlaceSelect/>
+                <MapInteractionsBar/>
                 <TimeSelect/>
+                <TimePlayer/>
                 <TimeSlider/>
-                <TimeSeriesModeSelect/>
             </ControlBar>
         );
     }
