@@ -1,5 +1,5 @@
 import { Variable } from './variable';
-import { isValidPlaceGroup, Place, PlaceGroup } from './place';
+import { PlaceGroup } from './place';
 import { TimeRange } from './timeSeries';
 import {
     assertArrayNotEmpty,
@@ -44,42 +44,6 @@ export function findDataset(datasets: Dataset[], selectedDatasetId: string | nul
 
 export function findDatasetVariable(dataset: Dataset, variableName: string | null): Variable | null {
     return (variableName && dataset.variables.find(variable => variable.name === variableName)) || null;
-}
-
-export function findDatasetPlace(dataset: Dataset, placeId: string | null): Place | null {
-    return (placeId && dataset.placeGroups && findPlaceInPlaceGroups(dataset.placeGroups, placeId)) || null;
-}
-
-export function findPlaceInPlaceGroups(placeGroups: PlaceGroup[], placeId: string | null): Place | null {
-    if (placeId) {
-        for (let placeGroup of placeGroups) {
-            const place = findPlaceInPlaceGroup(placeGroup, placeId);
-            if (place !== null) {
-                return place;
-            }
-        }
-    }
-    return null;
-}
-
-export function findPlaceInPlaceGroup(placeGroup: PlaceGroup, placeId: string | null): Place | null {
-    if (!placeId || !isValidPlaceGroup(placeGroup)) {
-        return null;
-    }
-    const place = placeGroup.features.find(place => place.id === placeId);
-    if (!!place) {
-        return place as Place;
-    }
-    let subPlaceGroups = placeGroup.placeGroups;
-    if (subPlaceGroups) {
-        for (let parentPlaceId in subPlaceGroups) {
-            const place = findPlaceInPlaceGroup(subPlaceGroups[parentPlaceId], placeId);
-            if (!!place) {
-                return place;
-            }
-        }
-    }
-    return null;
 }
 
 export function getDatasetTimeDimension(dataset: Dataset): TimeDimension {
