@@ -1,4 +1,5 @@
 import * as  GeoJSON from 'geojson';
+import { USER_PLACES_COLOR_NAMES } from '../config';
 
 
 /**
@@ -18,14 +19,45 @@ export interface PlaceGroup extends GeoJSON.FeatureCollection {
     placeGroups?: { [placeId: string]: PlaceGroup }; // placeGroups in placeGroups are not yet supported
 }
 
+/**
+ * Precomputed stuff.
+ */
+export interface PlaceInfo {
+    placeGroup: PlaceGroup;
+    place: Place;
+    label: string;
+    color: string;
+}
 
-export const LABEL_PROPERTY_NAMES = ['label', 'title', 'name', 'id'];
 
+export const DEFAULT_LABEL_PROPERTY_NAMES = ['label', 'LABEL', 'Label',
+                                             'title', 'TITLE', 'Title',
+                                             'name', 'NAME', 'Name',
+                                             'id', 'ID', 'Id'];
 
-export function getPlaceLabel(place: Place, labelPropNames: string []) {
+export function forEachPlace(placeGroups: PlaceGroup[], callback: (placeGroup: PlaceGroup, place: Place, label: string, color: string) => void) {
+    placeGroups.forEach(placeGroup => {
+        if (isValidPlaceGroup(placeGroup)) {
+            const labelPropNames = getPlaceGroupLabelPropertyNames(placeGroup);
+            placeGroup.features.forEach((place: Place) => {
+                callback(placeGroup, place, getPlaceLabel(place, labelPropNames), getPlaceColor(place));
+            });
+        }
+    });
+}
+
+function getPlaceGroupLabelPropertyNames(placeGroup: PlaceGroup): string[] {
+    const propertyMapping = placeGroup.propertyMapping;
+    if (propertyMapping && propertyMapping['label']) {
+        return [propertyMapping['label'], ...DEFAULT_LABEL_PROPERTY_NAMES];
+    }
+    return DEFAULT_LABEL_PROPERTY_NAMES;
+}
+
+function getPlaceLabel(place: Place, labelPropertyNames: string []) {
     if (place.properties) {
         let label;
-        for (let propName of labelPropNames) {
+        for (let propName of labelPropertyNames) {
             label = place.properties[propName];
             if (label) {
                 return label;
@@ -35,6 +67,59 @@ export function getPlaceLabel(place: Place, labelPropNames: string []) {
     return '' + place.id;
 }
 
+function getPlaceColor(place: Place) {
+    return place.properties && place.properties['color']
+           || USER_PLACES_COLOR_NAMES[getPlaceHash(place) % USER_PLACES_COLOR_NAMES.length]
+}
+
+function getPlaceHash(place: Place): number {
+    const id = place.id + '';
+    let hash = 0, i, c;
+    if (id.length === 0) {
+        return hash;
+    }
+    for (i = 0; i < id.length; i++) {
+        c = id.charCodeAt(i);
+        hash = ((hash << 5) - hash) + c;
+        hash |= 0; // Convert to 32bit integer
+    }
+    return hash;
+}
+
 export function isValidPlaceGroup(placeGroup: PlaceGroup): boolean {
     return !!placeGroup.features;
 }
+
+
+export function findPlaceInPlaceGroup(placeGroup: PlaceGroup, placeId: string | null): Place | null {
+    if (!placeId || !isValidPlaceGroup(placeGroup)) {
+        return null;
+    }
+    const place = placeGroup.features.find(place => place.id === placeId);
+    if (!!place) {
+        return place as Place;
+    }
+    let subPlaceGroups = placeGroup.placeGroups;
+    if (subPlaceGroups) {
+        for (let parentPlaceId in subPlaceGroups) {
+            const place = findPlaceInPlaceGroup(subPlaceGroups[parentPlaceId], placeId);
+            if (!!place) {
+                return place;
+            }
+        }
+    }
+    return null;
+}
+
+export function findPlaceInPlaceGroups(placeGroups: PlaceGroup[], placeId: string | null): Place | null {
+    if (placeId) {
+        for (let placeGroup of placeGroups) {
+            const place = findPlaceInPlaceGroup(placeGroup, placeId);
+            if (place !== null) {
+                return place;
+            }
+        }
+    }
+    return null;
+}
+
